pyproject.toml
[tool.poetry]
name = "sentinel"
version = "0.1.0"
description = ""
authors = ["vinicius <vyniciuss@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
loguru = "^0.7.2"
pyspark = "^3.5.1"
delta-spark = "^3.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.12.2"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.24"
mkdocstrings = "^0.25.1"
mkdocstrings-python = "^1.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=sentinel -vv"
post_test = "coverage html"